Refactor: Reorganize core grammar and expression components
This commit implements the initial phase of the codebase reorganization focused on the core grammar and expression functionalities.

Key changes include:
- Created a new directory structure under `janus/core` and other specified locations.
- Moved `progressive_grammar_system.py` to `janus/core/grammar.py`.
- Created `janus/core/expression.py`.
- Extracted the `Expression` and `Variable` classes from `janus/core/grammar.py` and moved them into `janus/core/expression.py`.
- Updated all import statements across the codebase to reflect these changes, ensuring absolute imports from the `janus` root.
- Updated the name of `tests/test_progressive_grammar.py` to `tests/test_grammar.py` and updated its internal references and patch paths.

This refactoring aims to improve modularity and clarity of the codebase by separating data structures (Expression, Variable) from the grammar logic.
