+++ maml_training_framework.py
@@ -25,7 +25,8 @@ import types # Ensure types is imported
 # Import existing Janus components
 from janus.ai_interpretability.environments import SymbolicDiscoveryEnv
 from hypothesis_policy_network import HypothesisNet, PPOTrainer
-from progressive_grammar_system import ProgressiveGrammar, Variable
+from janus.core.grammar import ProgressiveGrammar
+from janus.core.expression import Variable
 from janus.ai_interpretability.utils.math_utils import calculate_symbolic_accuracy, safe_env_reset # Added import
 import sympy as sp
