+++ tests/test_grammar.py
@@ -33,7 +33,8 @@ if 'sklearn' not in sys.modules:
     sys.modules['sklearn.preprocessing'] = sklearn_preproc_stub
 
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-from progressive_grammar_system import ProgressiveGrammar, Variable, Expression
+from janus.core.grammar import ProgressiveGrammar
+from janus.core.expression import Variable, Expression
 import sympy as sp
 import numpy as np
 
@@ -710,9 +711,9 @@ def test_generate_variable_name(grammar_and_vars):
 
 
 # Patching 'sklearn.decomposition.FastICA' as it's imported like `from sklearn.decomposition import FastICA`
-# Patching denoiser at the class level progressive_grammar_system.NoisyObservationProcessor
-@patch('progressive_grammar_system.FastICA')
-@patch('progressive_grammar_system.NoisyObservationProcessor.denoise')
+# Patching denoiser at the class level janus.core.grammar.NoisyObservationProcessor
+@patch('janus.core.grammar.FastICA')
+@patch('janus.core.grammar.NoisyObservationProcessor.denoise')
 def test_discover_variables_simple_run(mock_denoise, MockFastICA, grammar_and_vars):
     grammar, _, _, _, _, _ = grammar_and_vars
     grammar.variables = {}
@@ -763,9 +764,9 @@ def test_discover_variables_simple_run(mock_denoise, MockFastICA, grammar_and_va
         assert grammar.variables[var.name] == var
 
 
-@patch('progressive_grammar_system.FastICA')
+@patch('janus.core.grammar.FastICA')
 @patch.object(ProgressiveGrammar, '_analyze_component')
-@patch('progressive_grammar_system.NoisyObservationProcessor')
+@patch('janus.core.grammar.NoisyObservationProcessor')
 def test_discover_variables_controlled(MockNoisyObservationProcessor, mock_analyze_component, MockFastICA):
     # Create ProgressiveGrammar instance *after* mocks are in place
     grammar = ProgressiveGrammar()
@@ -786,7 +787,7 @@ def test_discover_variables_controlled(MockNoisyObservationProcessor, mock_analy
     mock_denoiser_instance.denoise.side_effect = lambda obs, epochs=50: obs
 
     # Alternative: If grammar.denoiser was already created *before* this test's patches apply (e.g. in fixture)
-    # then patching the class 'progressive_grammar_system.NoisyObservationProcessor' would not affect
+    # then patching the class 'janus.core.grammar.NoisyObservationProcessor' would not affect
     # the already existing grammar.denoiser instance.
     # The fixture `grammar_and_vars` creates a new ProgressiveGrammar instance for each test.
     # So, when ProgressiveGrammar() is called within the test's scope (or by the fixture for the test),
@@ -1093,7 +1094,7 @@ def test_mine_abstractions(mock_add_learned_function, grammar_and_vars):
 
 
 # --- Tests for NoisyObservationProcessor ---
-from progressive_grammar_system import NoisyObservationProcessor
+from janus.core.grammar import NoisyObservationProcessor
 from unittest.mock import PropertyMock
 
 def test_simple_denoise():
@@ -1125,10 +1126,10 @@ def test_simple_denoise():
     assert np.array_equal(denoised_small_15, obs_small_15_rows)
 
 
-@patch('progressive_grammar_system.StandardScaler')
-@patch('progressive_grammar_system.torch.FloatTensor')
-@patch('progressive_grammar_system.torch.optim.Adam')
-@patch('progressive_grammar_system.torch.nn.MSELoss')
+@patch('janus.core.grammar.StandardScaler')
+@patch('janus.core.grammar.torch.FloatTensor')
+@patch('janus.core.grammar.torch.optim.Adam')
+@patch('janus.core.grammar.torch.nn.MSELoss')
 @patch.object(NoisyObservationProcessor, 'build_autoencoder')
 def test_denoise_structure(mock_build_autoencoder, mock_loss, mock_adam, mock_float_tensor, MockStandardScaler):
     processor = NoisyObservationProcessor()
